Update get_dataset_PIDs.py
Improves how script gets PIDs of all datasets in the repository (by reducing the number of API calls).

If the user enters a URL with no dataverse alias or whose alias is the root dataverse's alias, the less resource intensive Search API is used to get all of the repository's PIDs.

Otherwise, if the user enters a URL of a dataverse that's not the root dataverse, the Native API's "get content" endpoints are used.