Scripts create and/or read from CSV files with more dataset info
Scripts create and/or read from CSV files with more dataset info. Instead of creating or reading from a text file of dataset PIDs, the scripts now create and/or read from a CSV file that includes each dataset's PID in canonical and URL form, and the name and alias of its owning dataverse.