Update get_dataset_PIDs.py
Adding help text about "Include subdataverses" checkbox. If the Dataverse collection URL entered is the "Root", all datasets will be found, so the checkbox is ignored for now.

This can be revisited if there's a need to get only the datasets that are children of the "Root" dataverse (and not within "subdataverses").