Update dataverse_repository_curation_assistant_functions.py
Closes #29

The problem wasn't with deaccessioned datasets, but with any search URLs that contains HTML encoding, like %20, instead of the human-readable value that it represents, like a ' ' (space) character.

For some reason, the expected results aren't returned when the query parameter contains HTML encoding, but works when the space character is used instead, like ' '.

So  when someone enters "https://dataverse.harvard.edu/dataverse/harvard?q=%22Arif,%20Mohd%22", before the get_params function gets the parameters of that search url to pass to the Requests get function, it needs to convert any HTML encoding in the query parameter to human-readable characters first.