Update get_dataset_metadata_of_all_installations.py
Using joblib library to use 4 of the computer's processor cores for the most time consuming parts of the script: paginating through an installation's Search API to get basic info about each dataset, and using the Native API endpoints for downloading the "Dataverse JSON" exports of each version of each file.

Also adds an argument to the function for downloading a dataset's "Dataverse JSON" export, to download the metadata of all versions of a dataset or the metadata of only the latest version.

Simplifies logic for creating CSV file that lists basic info about each dataset and whether or not the script was able to download its "Dataverse JSON" file.